videoflix_backend/movie_app/api/serializers.py
from rest_framework import serializers
from movie_app.models import Movie, Trailer

class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = '__all__'

class TrailerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Trailer
        fields = '__all__'       